remove glTF mesh utilities from core
