--- modules/core/src/categories/mesh/index.js
@@ -1,8 +1,0 @@
-export {getAccessorTypeFromSize, getComponentTypeFromArray} from './gltf-type-utils';
-
-export {
-  getGLTFAttribute,
-  getGLTFAccessors,
-  getGLTFIndices,
-  getGLTFAttributeMap
-} from './gltf-attribute-utils';
