+++ modules/draco/test/index.js
@@ -1,3 +1,4 @@
+import './gltf-attribute-utils.spec';
 import './draco-encoder.spec';
 import './draco-loader.spec';
 import './draco-compression-ratio.spec';
